requirements.txt
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.1.1
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.1
django-admin-rangefilter==0.6.3
django-allauth==0.42.0
django-extensions==2.2.5
django-pagseguro2==3.0.0
django-simple-email-confirmation==0.70
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
isort==5.5.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==8.0.0
progress==1.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
webencodings==0.5.1
Werkzeug==0.16.0
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
xmltodict==0.12.0
